renamed activity data
